actualizando archivos

+++ Backend-sistema-ventas/src/main/java/com/empresa/ventas/sistema_ventas/dto/response/DetalleVentaResponse.java
@@ -1,15 +1,16 @@
 package com.empresa.ventas.sistema_ventas.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class DetalleVentaResponse {
-    private Long id;
+    private Integer id;
     private String productoNombre;
     private String productoCodigo;
     private BigDecimal cantidad;
